apps/doc-processor/pyproject.toml
[project]
name = "doc-processor"
version = "0.1.0"
description = "Document processing service for PDF and image text extraction"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pymupdf>=1.24.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    # Face recognition
    "insightface>=0.7.3",
    "onnxruntime>=1.19.0",
    "opencv-python-headless>=4.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "pyright>=1.1.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
